requirements.txt
black==19.10b0
imagesize==1.2.0
tqdm==4.46.1
Pillow==7.1.2
opencv-python==4.2.0.34
seaborn==0.10.1
pandas==1.0.4
numpy==1.18.4
scikit-image==0.17.2
torch==1.5.0
torchvision==0.6.0
